Initial commit
This project was separated from a C++ project nuBallCubes.
The scripts moved to this one are a Python scripts for making
projections (gates), while C++ code is for root to HDF5 event file
conversion.

In work directory, there are older version of nucubes (including other
detectors gating), nusquares for angular correlations, and test Julia
script (unfortunately not faster than python).